ai/src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "Rakesh Kanneeswaran",email = "rakikanneeswaran@gmail.com"}
]
readme = "README.md"

requires-python = ">=3.11,<4.0"

dependencies = [
    "black (>=25.1.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "whisper (>=1.1.10,<2.0.0)",
    "pyannote-audio (>=3.3.2,<4.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "boto3 (>=1.40.5,<2.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-core (>=0.3.74,<0.4.0)",
    "langchain-groq (>=0.3.7,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
